RMX1901: Use TARGET_SCREEN_DENSITY
This variable is used to auto-populate the ro.sf.lcd_density as
well as selecting density bucket for resources in recovery.

Bug: 118676832
Test: adb shell getprop ro.sf.lcd_density
Test: boot into recovery
Change-Id: Ib383d18b61e7a2058a69ae26136aa053fdba4df2
Signed-off-by: kaderbava <ksbava7325@gmail.com>
RMX1921: Use TARGET_SCREEN_DENSITY

This commit adds the use of TARGET_SCREEN_DENSITY variable in BoardConfig.mk. The variable is set to 420, which will auto-populate the ro.sf.lcd_density and select the density bucket for resources in recovery. This change is made to address Bug 118676832.

Tested by checking the value of ro.sf.lcd_density using "adb shell getprop" command and booting into recovery mode.

Change-Id: Ib383d18b61e7a2058a69ae26136aa053fdba4df2